backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.36
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pydantic[email]==2.10.3
python-multipart==0.0.18
